% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSID_Package.Documentation_function.R
\docType{package}
\name{easyPSID-package}
\alias{easyPSID}
\alias{easyPSID-package}
\title{easyPSID: Reading, Formatting, and Organizing the Panel Study of Income Dynamics (PSID)}
\description{
Provides various functions for reading and preparing the Panel Study of Income Dynamics (PSID) for longitudinal analysis, including functions that read the PSID's fixed width format files directly into R, rename all of the PSID's longitudinal variables so that recurring variables have consistent names across years, simplify assembling longitudinal datasets from cross sections of the PSID Family Files, and export the resulting PSID files into file formats common among other statistical programming languages ('SAS', 'STATA', and 'SPSS').
}
\details{
This package is designed for use with the PSID's packaged data Family Files, available at https://simba.isr.umich.edu/. See easyPSID's readme at https://github.com/BrianAronson/easyPSID/blob/master/README.md for a more detailed overview and guide for using this package.
}
\author{
\strong{Maintainer}: Brian Aronson \email{bdaronson@gmail.com}

}
