% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10find_desc.R
\name{find_description}
\alias{find_description}
\title{Find description of PSID variable}
\usage{
find_description(variables)
}
\arguments{
\item{variables}{Variable names to look up (as individual string or vector of strings)}
}
\description{
Finds the descriptions of selected PSID variables.
}
\examples{
find_description(variables=c("V2","V30"))
}
\keyword{PSID}
