% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8find_name.R
\name{find_name}
\alias{find_name}
\title{Lookup new longitudinal variables names}
\usage{
find_name(variable, var_year = TRUE)
}
\arguments{
\item{variable}{Variable to look up}

\item{var_year}{Report year when renamed variable first entered the PSID dataset (default=TRUE)}
}
\description{
Finds the new name of any longitudinal variable in the PSID Family Files or Individual files following implementation of the rename_fam_vars and rename_ind_vars functions.
}
\examples{
find_name(variable="V1244",var_year=FALSE)
}
\keyword{PSID}
