% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9find_years.R
\name{find_years}
\alias{find_years}
\title{Find years of PSID variable}
\usage{
find_years(variable, var_names = TRUE)
}
\arguments{
\item{variable}{Variable name to look up}

\item{var_names}{Report names for longitudinal PSID variable across years (default=TRUE)}
}
\description{
Finds the years and corresponding variable names for any longitudinal PSID variable in the PSID Family Files and Individual file.
}
\examples{
find_years(variable="V3",var_names=FALSE)
}
\keyword{PSID}
